pyproject.toml
[project]
name = "lakebase-mcp"
version = "0.2.0"
description = "Databricks Lakebase MCP Server for AI Agents — Autoscaling-Aware"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.9.0",
    "psycopg[binary]>=3.2.0",
    "psycopg-pool>=3.2.0",
    "pydantic>=2.0.0",
    "databricks-sdk>=0.40.0",
    "httpx>=0.27.0",
]

[project.scripts]
lakebase-mcp = "server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]
